requirements.txt
asgiref==3.5.2
attrs==22.1.0
boto3==1.26.151
botocore==1.29.151
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
django-cors-headers==3.13.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.1
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.3.0
platformdirs==2.5.4
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==7.0.0
pytz==2022.6
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.10.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
twilio==7.16.1
uritemplate==4.1.1
urllib3==1.26.13
zeep==4.2.1
